requirements.txt
pytest
pytest-playwright
playwright
pytest-html-reporter
allure-pytest